modification in the query handler to implement a facade design pattern
