+++ src/main/db/add.js
@@ -1,0 +1,9 @@
+import { db } from '../db/databaseStart.js';
+const collection = 'MSM-Scenario';
+
+export class Add {
+    static async add(data) {
+        const scenario = await db.collection(collection).add(data);
+        return scenario.id;
+    }
+}
