+++ src/main/db/set.js
@@ -1,0 +1,8 @@
+import { db } from '../db/databaseStart.js';
+const collection = 'MSM-Scenario';
+
+export class Set {
+    static async set(docId, newData) {
+        await db.collection(collection).doc(docId).set(newData);
+    }
+}
